requirements.txt
base58==2.1.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
idna==3.3
pyelftools==0.28
requests==2.27.1
tqdm==4.62.3
urllib3==1.26.8
